+++ packages/protocol/src/protocol/MessageExchange.ts
@@ -646,8 +646,16 @@ export class MessageExchange {
         return this.#timedInteractionTimer !== undefined && !this.#timedInteractionTimer.isRunning;
     }
 
-    async close() {
-        if (this.#closeTimer !== undefined) return; // close was already called
+    async close(force = false) {
+        if (this.#closeTimer !== undefined) {
+            if (force) {
+                // Force close does not wait any longer
+                this.#closeTimer.stop();
+                return this.#close();
+            }
+            // close was already called, so let retries happen because close not forced
+            return;
+        }
         this.#isClosing = true;
 
         if (this.#receivedMessageToAck !== undefined) {
@@ -659,7 +667,11 @@ export class MessageExchange {
             } catch (error) {
                 logger.error("An error happened when closing the exchange", error);
             }
-        } else if (this.#sentMessageToAck === undefined) {
+            if (force) {
+                // We have sent the Ack, so close here, no retries because close is forced
+                return this.#close();
+            }
+        } else if (this.#sentMessageToAck === undefined || force) {
             // No message left that we need to ack and no sent message left that waits for an ack, close directly
             return this.#close();
         }