+++ CHANGELOG.md
@@ -14,6 +14,9 @@ The main work (all changes without a GitHub username in brackets in the below li
 - @project-chip/matter.js
     - Fix: Allows more cases when checking if a device is battery powered to address real world devices
 
+-   @matter/protocol
+    - Fix: Added force closing of exchanges on shutdown of the node
+
 ## 0.12.0 (2025-01-23)
 
 -   @matter/general